# public/20191114 cissukl.swagger.json
{
  "info": {
    "description": "Popis RESTful API",
    "version": "1.0.0",
    "title": "Číselníky SÚKL"
  },
  "host": "localhost:8000",
  "basePath": "/cissuklapi/v1",
  "consumes": [
    "application/json"
  ],
  "paths": {
...
    "/stavyregistrace": {
      "get": {
        "tags": [
          "Ciselniky"
        ],
        "description": "Vrací seznam informací o stavech registrace",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Kód stavu registrace",
            "required": false,
            "type": "string",
            "enum": [
              "kod_stav_registrace"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/StavyRegistrace"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/stavyregistrace/{kod_stav_registrace}": {
      "get": {
        "tags": [
          "Ciselniky"
        ],
        "description": "Vrací informace o stavu registrace pro daný kód",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "kod_stav_registrace",
            "in": "path",
            "description": "Kód stavu registrace",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "description": "Operace proběhla bez chyby",
            "schema": {
              "$ref": "#/definitions/StavyRegistrace"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/indikacniskupiny": {
      "get": {
        "tags": [
          "Ciselniky"
        ],
        "description": "Vrací seznam informací o indikačních skupinách",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Kód indikační skupiny",
            "required": false,
            "type": "string",
            "enum": [
              "kod_indikacni_skupina"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/IndikacniSkupiny"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/indikacniskupiny/{kod_indikacni_skupina}": {
      "get": {
        "tags": [
          "Ciselniky"
        ],
        "description": "Vrací informace o indikační skupině pro daný kód",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "kod_indikacni_skupina",
            "in": "path",
            "description": "Kód indikační skupiny",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/IndikacniSkupiny"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/atcskupiny": {
      "get": {
        "tags": [
          "Ciselniky"
        ],
        "description": "Vrací seznam informací o ATC skupinách",
        "produces": [
          "text"
        ],
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Kód ATC skupiny",
            "required": false,
            "type": "string",
            "enum": [
              "kod_atc_skupina"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/AtcSkupiny"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/atcskupiny/{kod_atc_skupina}": {
      "get": {
        "tags": [
          "Ciselniky"
        ],
        "description": "Vrací informace a ATC skupině pro daný kód",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "kod_atc_skupina",
            "in": "path",
            "description": "Kód ATC skupiny",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/AtcSkupiny"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/lecivepripravky": {
      "get": {
        "tags": [
          "Lecive pripravky"
        ],
        "description": "Vrací seznam informací o léčivých přípravcích",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Kód SÚKL léčivého přípravku",
            "required": false,
            "type": "string",
            "enum": [
              "kod_sukl"
            ]
          },
          {
            "name": "stavy_registrace",
            "in": "query",
            "description": "Vybrané stavy registrace",
            "required": false,
            "type": "string",
            "enum": [
              "stavy_registrace_scau"
            ]
          },
          {
            "name": "je_regulovany",
            "in": "query",
            "description": "Příznak, zda je léčivý přípravek regulovaný",
            "required": false,
            "type": "string",
            "enum": [
              "A",
              "N"
            ]
          },
          {
            "name": "platnost_od",
            "in": "query",
            "description": "Datum, od kdy jsou informace o léčivém přípravku platné",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit, přednastavená hodnota limit=20",
            "required": false,
            "type": "number"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset, přednastavená hodnota offset=0",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/LecivePripravky"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
    "/lecivepripravky/{kod_sukl}": {
      "get": {
        "tags": [
          "Lecive pripravky"
        ],
        "description": "Vrací seznam informací o léčivých přípravcích",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "kod_sukl",
            "in": "path",
            "description": "Kód SÚKL léčivého přípravku",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/LecivePripravky"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
...
    "/ucinnelatky": {
      "get": {
        "tags": [
          "Ciselniky"
        ],
        "description": "Vrací seznam informací o účinných látkách",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "fields",
            "in": "query",
            "description": "Kód účinné látky",
            "required": false,
            "type": "string",
            "enum": [
              "kod_ucinna_latka"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Operace proběhla bez chyby",
            "headers": {
              "x-total-count": {
                "description": "Celkový počet nalezených záznamů",
                "type": "integer"
              }
            },
            "schema": {
              "$ref": "#/definitions/UcinneLatky"
            }
          },
          "400": {
            "description": "Pro dané URL není služba implementována",
            "schema": {
              "$ref": "#/definitions/Error400"
            }
          },
          "404": {
            "description": "Nenalezeny žádné záznamy",
            "schema": {
              "$ref": "#/definitions/Error404"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Error400": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Pro dané URL není služba implementována",
          "maxLength": 256
        }
      }
    },
    "Error404": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Nenalezeny žádné záznamy",
          "maxLength": 256
        }
      }
    },
...
    "LecivePripravky": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/LecivyPripravek"
      }
    },
...
    "StavyRegistrace": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/StavRegistrace"
      }
    },
    "StavRegistrace": {
      "type": "object",
      "properties": {
        "KOD_STAV_REGISTRACE": {
          "type": "string",
          "example": "R",
          "description": "Kód stavu registrace",
          "maxLength": 2
        },
        "NAZEV": {
          "type": "string",
          "example": "registrovaný léčivý přípravek",
          "description": "Název stavu registrace česky",
          "maxLength": 300
        }
      }
    },
    "IndikacniSkupiny": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/IndikacniSkupina"
      }
    },
    "IndikacniSkupina": {
      "type": "object",
      "properties": {
        "KOD_INDIKACNI_SKUPINA": {
          "type": "number",
          "example": 1,
          "description": "Kód indikační skupiny"
        },
        "NAZEV": {
          "type": "string",
          "example": "Anaesthetica (lokální)",
          "description": "Název indikační skupiny česky",
          "maxLength": 58
        }
      }
    },
    "AtcSkupiny": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/AtcSkupina"
      }
    },
    "AtcSkupina": {
      "type": "object",
      "properties": {
        "KOD_ATC_SKUPINA": {
          "type": "string",
          "example": "A01AA02",
          "description": "Kód ATC skupiny",
          "maxLength": 7
        },
        "NAZEV": {
          "type": "string",
          "example": "Monofluorofosforečnan sodný",
          "description": "Název ATC skupiny česky",
          "maxLength": 200
        },
        "NAZEV_EN": {
          "type": "string",
          "example": "sodium monofluorophosphate",
          "description": "Název ATC skupiny anglicky",
          "maxLength": 200
        },
        "JE_SUKL_ATC_SKUPINA": {
          "type": "string",
          "example": "N",
          "description": "příznak, zda ATC skupina je definovaná na SÚKL, hodnoty \"A\" - ano, \"N\" - ne",
          "maxLength": 1
        }
      }
    },
    "UcinneLatky": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/UcinnaLatka"
      }
    },
...
  },
  "produces": [
    "application/json"
  ],
  "swagger": "2.0"
}